==> 08-Predict-SalePrice.ipynb <==
# Regression with all transactions included

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import matplotlib.pyplot as plt
import seaborn as sns
sns.set_style('whitegrid')
import xgboost as xgb
import warnings
warnings.simplefilter(action='ignore')
from scipy.optimize import minimize
from scipy.optimize import fmin
import math
from bayes_opt import BayesianOptimization

# Loading Data 

Tot = pd.read_csv('austin_housing_data.csv')
print('Number of Transactions: ' + str(len(Tot)))

# This step randomly sorts the transactions
Tot.sample(frac=1)
Tot.columns

# Eliminate Columns
colToElim = ['Unnamed: 0', 'city', 'latitude', 'longitude', 'pricePerSqFt', 'state', 'streetLine', 'zip',  'idx', 
             'Population Density', 'Address', 'Number', 'daysOnMarket']
Tot.drop(colToElim, axis=1, inplace=True)
Tot.columns

aTot = Tot.copy()
# Delete columns for testing purposes
colToDel = ['daysAftSold', 'daysBtwBldAndSold', 'Crime Index', 'park_index', 'schools_index', 'event_index', 'fastfood_index', 
            'Crime Index Scaled', '2015 Appraisal', '2019 Appraisal' , '5-year Growth', 'event_index_scaled', 'fastfood_index_scaled',
           'park_index_scaled', 'school_index_scaled']
aTot.drop(colToDel, axis=1, inplace=True)
colBefZip = aTot.shape[1]
print("Number of Columns: " + str(colBefZip))
aTot.columns

# Linear Regression to predict sale price

# Linear Regression library
from scipy.stats import skew
from sklearn.linear_model import Ridge, RidgeCV, ElasticNet, LassoCV, LassoLarsCV
from sklearn.model_selection import cross_val_score
from sklearn.metrics import mean_squared_error


# Objective function definition
X_train = []
y_train = []

def target(x):
    model = Ridge(alpha = x)
    rmse= np.sqrt(-cross_val_score(model, X_train, y_train, scoring="neg_mean_squared_error", cv = 3)).mean()     
    return(-rmse)    

# Select 80% of the data for training
nTot = len(aTot)
msk = np.random.rand(nTot) < 0.8

a = aTot[msk]
na = len(a)
ytrain = a['salePrice'].to_frame()
a.drop(['salePrice'], axis=1, inplace=True)

# The remaining 20% is used for testing
b = aTot[~msk]
nb = len(b)
ytest = b['salePrice'].to_frame()
b.drop(['salePrice'], axis=1, inplace=True)


c = pd.concat((a, b), sort=False).reset_index(drop=True)
numeric_feats = c.dtypes[c.dtypes != "object"].index
skewed_feats = c[numeric_feats].apply(lambda x: skew(x.dropna())) #compute skewness
skewed_feats = skewed_feats[skewed_feats > 0.75]
skewed_feats = skewed_feats.index
c[skewed_feats] = np.log1p(c[skewed_feats])


X_train = c[:na]   
X_test = c[na:]

y_train = np.log1p(ytrain) 
y_test = np.log1p(ytest) 


# define the optimizer
pbounds = {'x': (0, 200)}    
optimizer = BayesianOptimization(f=target, pbounds=pbounds, verbose=0, random_state=1,)
optimizer.maximize(init_points=2,n_iter=50,)
print(optimizer.max)   
sol = optimizer.max['params']['x']
model_ridge = Ridge(alpha = sol)
model_ridge.fit(X_train,y_train)  

# Sorting between best and worst zip codes
dfCoef = pd.DataFrame(np.transpose(model_ridge.coef_), index =X_train.columns, columns =['Coef']) 
print(dfCoef.sort_values(by=['Coef'], ascending = False))

# Mean Square Error for the Train set

Ypredict = model_ridge.predict(X_train)
predictError = y_train -  Ypredict
MSE_Train = mean_squared_error(y_train,Ypredict)
print('Train MSE = ' + str(MSE_Train))

# Mean Square Error for the Testing set

Ypredict = model_ridge.predict(X_test)
predictError = y_test -  Ypredict
MSE_Test = mean_squared_error(y_test,Ypredict)
print('Test MSE = ' + str(MSE_Test))

# Here we add one atribute at a time, to measure its effect

results = {'Init' : [MSE_Train, MSE_Test]}
for addFeature in colToDel :
    print(addFeature)
    addTot = aTot.copy()
    addTot[addFeature] = Tot[addFeature]
    
    # Select 80% of the data for training
    nTot = len(addTot)
    msk = np.random.rand(nTot) < 0.8

    a = addTot[msk]
    na = len(a)
    ytrain = a['salePrice'].to_frame()
    a.drop(['salePrice'], axis=1, inplace=True)

    # The remaining 20% is used for testing
    b = addTot[~msk]
    nb = len(b)
    ytest = b['salePrice'].to_frame()
    b.drop(['salePrice'], axis=1, inplace=True)


    c = pd.concat((a, b), sort=False).reset_index(drop=True)
    numeric_feats = c.dtypes[c.dtypes != "object"].index
    skewed_feats = c[numeric_feats].apply(lambda x: skew(x.dropna())) #compute skewness
    skewed_feats = skewed_feats[skewed_feats > 0.75]
    skewed_feats = skewed_feats.index
    c[skewed_feats] = np.log1p(c[skewed_feats])


    X_train = c[:na]   
    X_test = c[na:]

    y_train = np.log1p(ytrain) 
    y_test = np.log1p(ytest) 


    # define the optimizer
    pbounds = {'x': (0, 200)}
    optimizer = BayesianOptimization(f=target, pbounds=pbounds, verbose=0, random_state=1,)
    optimizer.maximize(init_points=2,n_iter=50,)
    print(optimizer.max)   
    sol = optimizer.max['params']['x']
    model_ridge = Ridge(alpha = sol)
    model_ridge.fit(X_train,y_train)      
    
    Ypredict = model_ridge.predict(X_train)
    predictError = y_train -  Ypredict
    MSE_Train = mean_squared_error(y_train,Ypredict)
    print('Train MSE = ' + str(MSE_Train))
    Ypredict = model_ridge.predict(X_test)
    predictError = y_test -  Ypredict
    MSE_Test = mean_squared_error(y_test,Ypredict)
    print('Test MSE = ' + str(MSE_Test))
    results[addFeature] = [MSE_Train, MSE_Test]

# Printing the results of how each feature impacts the sale price prediction

results

# Calculation of MSE percent improvement - 

TrMSE = results['Init'][0]
TeMSE = results['Init'][1]
invTrMSE = 100/TrMSE
invTeMSE = 100/TeMSE
resPer = {}
for (key, value) in results.items() :
    TrPer = (value[0] - TrMSE) * invTrMSE
    TePer = (value[1] - TeMSE) * invTeMSE    
    resPer[key] = [TrPer, TePer]
resPer    

# Now we repeat the calculation but includying days after sold 'daysAftSold' is the most influential from previous calculations

aTot = Tot.copy()
# Delete columns for testing purposes
colToDel = ['daysBtwBldAndSold', 'Crime Index', 'park_index', 'schools_index', 'event_index', 'fastfood_index', 
            'Crime Index Scaled', '2015 Appraisal', '2019 Appraisal' , '5-year Growth', 'event_index_scaled', 'fastfood_index_scaled',
           'park_index_scaled', 'school_index_scaled']
aTot.drop(colToDel, axis=1, inplace=True)
colBefZip = aTot.shape[1]
print("Number of Columns: " + str(colBefZip))
aTot.columns

# Select 80% of the data for training

nTot = len(aTot)
msk = np.random.rand(nTot) < 0.8

a = aTot[msk]
na = len(a)
ytrain = a['salePrice'].to_frame()
a.drop(['salePrice'], axis=1, inplace=True)

# The remaining 20% is used for testing
b = aTot[~msk]
nb = len(b)
ytest = b['salePrice'].to_frame()
b.drop(['salePrice'], axis=1, inplace=True)


c = pd.concat((a, b), sort=False).reset_index(drop=True)
numeric_feats = c.dtypes[c.dtypes != "object"].index
skewed_feats = c[numeric_feats].apply(lambda x: skew(x.dropna())) #compute skewness
skewed_feats = skewed_feats[skewed_feats > 0.75]
skewed_feats = skewed_feats.index
c[skewed_feats] = np.log1p(c[skewed_feats])


X_train = c[:na]   
X_test = c[na:]

y_train = np.log1p(ytrain) 
y_test = np.log1p(ytest) 


# define the optimizer
pbounds = {'x': (0, 200)}    
optimizer = BayesianOptimization(f=target, pbounds=pbounds, verbose=0, random_state=1,)
optimizer.maximize(init_points=2,n_iter=50,)
print(optimizer.max)   
sol = optimizer.max['params']['x']
model_ridge = Ridge(alpha = sol)
model_ridge.fit(X_train,y_train)  

# Sorting between best and worst zip codes

dfCoef = pd.DataFrame(np.transpose(model_ridge.coef_), index =X_train.columns, columns =['Coef']) 
print(dfCoef.sort_values(by=['Coef'], ascending = False))

# Mean Square Error for the Train set
Ypredict = model_ridge.predict(X_train)
predictError = y_train -  Ypredict
MSE_Train = mean_squared_error(y_train,Ypredict)
print('Train MSE = ' + str(MSE_Train))

# Mean Square Error for the Testing set
Ypredict = model_ridge.predict(X_test)
predictError = y_test -  Ypredict
MSE_Test = mean_squared_error(y_test,Ypredict)
print('Test MSE = ' + str(MSE_Test))

# Here we add one atribute at a time, to measure its effect
results = {'Init' : [MSE_Train, MSE_Test]}
for addFeature in colToDel :
    print(addFeature)
    addTot = aTot.copy()
    addTot[addFeature] = Tot[addFeature]
    
    # Select 80% of the data for training
    nTot = len(addTot)
    msk = np.random.rand(nTot) < 0.8

    a = addTot[msk]
    na = len(a)
    ytrain = a['salePrice'].to_frame()
    a.drop(['salePrice'], axis=1, inplace=True)

    # The remaining 20% is used for testing
    b = addTot[~msk]
    nb = len(b)
    ytest = b['salePrice'].to_frame()
    b.drop(['salePrice'], axis=1, inplace=True)


    c = pd.concat((a, b), sort=False).reset_index(drop=True)
    numeric_feats = c.dtypes[c.dtypes != "object"].index
    skewed_feats = c[numeric_feats].apply(lambda x: skew(x.dropna())) #compute skewness
    skewed_feats = skewed_feats[skewed_feats > 0.75]
    skewed_feats = skewed_feats.index
    c[skewed_feats] = np.log1p(c[skewed_feats])


    X_train = c[:na]   
    X_test = c[na:]

    y_train = np.log1p(ytrain) 
    y_test = np.log1p(ytest) 


    # define the optimizer
    pbounds = {'x': (0, 200)}
    optimizer = BayesianOptimization(f=target, pbounds=pbounds, verbose=0, random_state=1,)
    optimizer.maximize(init_points=2,n_iter=50,)
    print(optimizer.max)   
    sol = optimizer.max['params']['x']
    model_ridge = Ridge(alpha = sol)
    model_ridge.fit(X_train,y_train)      
    
    Ypredict = model_ridge.predict(X_train)
    predictError = y_train -  Ypredict
    MSE_Train = mean_squared_error(y_train,Ypredict)
    print('Train MSE = ' + str(MSE_Train))
    Ypredict = model_ridge.predict(X_test)
    predictError = y_test -  Ypredict
    MSE_Test = mean_squared_error(y_test,Ypredict)
    print('Test MSE = ' + str(MSE_Test))
    results[addFeature] = [MSE_Train, MSE_Test]

# Printing the results
results

# Calculation of MSE percent improvement - 
TrMSE = results['Init'][0]
TeMSE = results['Init'][1]
invTrMSE = 100/TrMSE
invTeMSE = 100/TeMSE
resPer = {}
for (key, value) in results.items() :
    TrPer = (value[0] - TrMSE) * invTrMSE
    TePer = (value[1] - TeMSE) * invTeMSE    
    resPer[key] = [TrPer, TePer]
resPer    

# Repat the calculation but includying Crime Index as the most influential from last calculations

aTot = Tot.copy()
# Eliminate columns with no valuable info (now we include )
colToDel = ['daysBtwBldAndSold', 'park_index', 'schools_index', 'event_index', 'fastfood_index', 
            'Crime Index Scaled', '2015 Appraisal', '2019 Appraisal' , '5-year Growth', 'event_index_scaled', 'fastfood_index_scaled',
           'park_index_scaled', 'school_index_scaled']
aTot.drop(colToDel, axis=1, inplace=True)
colBefZip = aTot.shape[1]
print("Number of Columns: " + str(colBefZip))
print(aTot.dtypes)

# Objective function definition
X_train = []
y_train = []

def target(x):
    model = Ridge(alpha = x)
    rmse= np.sqrt(-cross_val_score(model, X_train, y_train, scoring="neg_mean_squared_error", cv = 3)).mean()     
    return(-rmse)    

# Select 80% of the data for training
nTot = len(aTot)
msk = np.random.rand(nTot) < 0.8

a = aTot[msk]
na = len(a)
ytrain = a['salePrice'].to_frame()
a.drop(['salePrice'], axis=1, inplace=True)

# The remaining 20% is used for testing
b = aTot[~msk]
nb = len(b)
ytest = b['salePrice'].to_frame()
b.drop(['salePrice'], axis=1, inplace=True)


c = pd.concat((a, b), sort=False).reset_index(drop=True)
numeric_feats = c.dtypes[c.dtypes != "object"].index
skewed_feats = c[numeric_feats].apply(lambda x: skew(x.dropna())) #compute skewness
skewed_feats = skewed_feats[skewed_feats > 0.75]
skewed_feats = skewed_feats.index
c[skewed_feats] = np.log1p(c[skewed_feats])


X_train = c[:na]   
X_test = c[na:]

y_train = np.log1p(ytrain) 
y_test = np.log1p(ytest) 


# define the optimizer
pbounds = {'x': (0, 200)}    
optimizer = BayesianOptimization(f=target, pbounds=pbounds, verbose=0, random_state=1,)
optimizer.maximize(init_points=2,n_iter=50,)
print(optimizer.max)   
sol = optimizer.max['params']['x']
model_ridge = Ridge(alpha = sol)
model_ridge.fit(X_train,y_train)  

# Sorting between best and worst zip codes
dfCoef = pd.DataFrame(np.transpose(model_ridge.coef_), index =X_train.columns, columns =['Coef']) 
print(dfCoef.sort_values(by=['Coef'], ascending = False))

# Mean Square Error for the Train set
Ypredict = model_ridge.predict(X_train)
predictError = y_train -  Ypredict
MSE_Train = mean_squared_error(y_train,Ypredict)
print('Train MSE = ' + str(MSE_Train))

# Mean Square Error for the Testing set
Ypredict = model_ridge.predict(X_test)
predictError = y_test -  Ypredict
MSE_Test = mean_squared_error(y_test,Ypredict)
print('Test MSE = ' + str(MSE_Test))

# Here we add one atribute at a time, to measure its effect
results = {'Init' : [MSE_Train, MSE_Test]}
for addFeature in colToDel :
    print(addFeature)
    addTot = aTot.copy()
    addTot[addFeature] = Tot[addFeature]
    
    # Select 80% of the data for training
    nTot = len(addTot)
    msk = np.random.rand(nTot) < 0.8

    a = addTot[msk]
    na = len(a)
    ytrain = a['salePrice'].to_frame()
    a.drop(['salePrice'], axis=1, inplace=True)

    # The remaining 20% is used for testing
    b = addTot[~msk]
    nb = len(b)
    ytest = b['salePrice'].to_frame()
    b.drop(['salePrice'], axis=1, inplace=True)


    c = pd.concat((a, b), sort=False).reset_index(drop=True)
    numeric_feats = c.dtypes[c.dtypes != "object"].index
    skewed_feats = c[numeric_feats].apply(lambda x: skew(x.dropna())) #compute skewness
    skewed_feats = skewed_feats[skewed_feats > 0.75]
    skewed_feats = skewed_feats.index
    c[skewed_feats] = np.log1p(c[skewed_feats])


    X_train = c[:na]   
    X_test = c[na:]

    y_train = np.log1p(ytrain) 
    y_test = np.log1p(ytest) 


    # define the optimizer
    pbounds = {'x': (0, 200)}
    optimizer = BayesianOptimization(f=target, pbounds=pbounds, verbose=0, random_state=1,)
    optimizer.maximize(init_points=2,n_iter=50,)
    print(optimizer.max)   
    sol = optimizer.max['params']['x']
    model_ridge = Ridge(alpha = sol)
    model_ridge.fit(X_train,y_train)      
    
    Ypredict = model_ridge.predict(X_train)
    predictError = y_train -  Ypredict
    MSE_Train = mean_squared_error(y_train,Ypredict)
    print('Train MSE = ' + str(MSE_Train))
    Ypredict = model_ridge.predict(X_test)
    predictError = y_test -  Ypredict
    MSE_Test = mean_squared_error(y_test,Ypredict)
    print('Test MSE = ' + str(MSE_Test))
    results[addFeature] = [MSE_Train, MSE_Test]
